Predictive Modeling/requirements/requirements.txt
# Core Data Science Libraries
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0

# Deep Learning
tensorflow>=2.8.0
keras>=2.8.0
torch>=1.12.0
torchvision>=0.13.0

# Time Series Analysis
prophet>=1.1.0
statsmodels>=0.13.0
pmdarima>=1.8.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
bokeh>=2.4.0

# Jupyter and Interactive Development
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Data Processing and Utilities
joblib>=1.1.0
tqdm>=4.62.0
click>=8.0.0
pyyaml>=6.0

# Testing and Development
pytest>=6.2.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0

# Additional ML Libraries
catboost>=1.0.0
optuna>=2.10.0
shap>=0.40.0
eli5>=0.13.0

# Database and Data Sources
sqlalchemy>=1.4.0
requests>=2.27.0
beautifulsoup4>=4.10.0

# Configuration and Logging
python-dotenv>=0.19.0
loguru>=0.6.0

# Optional: R Integration (if using rpy2)
# rpy2>=3.4.0
